requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
filelock==3.12.0
frozenlist==1.3.3
idna==3.4
joblib==1.2.0
minimalkv==1.6.1
multidict==6.0.4
numpy==1.24.3
openai==0.27.4
pandas==2.0.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
requests-file==1.5.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
startupradar-transformers @ git+https://github.com/startupradar/transformers.git@885252e95931ee4ae447460caef19e90fc3a9cea
threadpoolctl==3.1.0
tldextract==3.4.0
tqdm==4.65.0
tzdata==2023.3
uritools==4.0.1
urllib3==1.26.15
yarl==1.9.1
